+++ src/main/scala/com/rawlabs/das/databricks/DASDatabricksConnection.scala
@@ -12,24 +12,108 @@
 
 package com.rawlabs.das.databricks
 
+import java.net.UnknownHostException
+
+import scala.annotation.tailrec
 import scala.jdk.CollectionConverters.IterableHasAsScala
 
 import com.databricks.sdk.WorkspaceClient
-import com.databricks.sdk.core.DatabricksConfig
-import com.databricks.sdk.service.catalog.{ListTablesRequest, TableInfo}
+import com.databricks.sdk.core.error.platform.{BadRequest, NotFound, PermissionDenied, Unauthenticated}
+import com.databricks.sdk.core.{DatabricksConfig, DatabricksError, DatabricksException}
+import com.databricks.sdk.service.catalog.ListTablesRequest
+import com.databricks.sdk.service.sql.{ExecuteStatementRequest, StatementResponse, StatementState}
+import com.rawlabs.das.sdk.{
+  DASSdkInvalidArgumentException,
+  DASSdkPermissionDeniedException,
+  DASSdkUnauthenticatedException
+}
+import com.typesafe.scalalogging.StrictLogging
 
-class DASDatabricksConnection(options: Map[String, String]) {
+class DASDatabricksConnection(options: Map[String, String]) extends StrictLogging {
 
-  private val host: String = options.getOrElse("host", throw new IllegalArgumentException("Host is required"))
-  private val token: String = options.getOrElse("token", throw new IllegalArgumentException("Token is required"))
-  private val catalog: String = options.getOrElse("catalog", throw new IllegalArgumentException("Catalog is required"))
-  private val schema: String = options.getOrElse("schema", throw new IllegalArgumentException("Schema is required"))
+  private val host: String = options.getOrElse("host", throw new DASSdkInvalidArgumentException("Host is required"))
+  private val token: String = options.getOrElse("token", throw new DASSdkInvalidArgumentException("Token is required"))
+  private val catalog: String =
+    options.getOrElse("catalog", throw new DASSdkInvalidArgumentException("Catalog is required"))
+  private val schema: String =
+    options.getOrElse("schema", throw new DASSdkInvalidArgumentException("Schema is required"))
+  private val warehouse: String =
+    options.getOrElse("warehouse", throw new DASSdkInvalidArgumentException("Warehouse ID is required"))
   private val config = new DatabricksConfig().setHost(host).setToken(token)
-  private val w = new WorkspaceClient(config)
+  private val databricksClient = new WorkspaceClient(config)
 
-  def tableDefinitions(): Seq[TableInfo] = {
+  def dasTables(): Iterable[DASDatabricksTable] = {
     val req = new ListTablesRequest().setCatalogName(catalog).setSchemaName(schema)
-    val tables = w.tables().list(req).asScala
-    tables.toList
+    withDatabricksException {
+      databricksClient
+        .tables()
+        .list(req)
+        .asScala
+        .map(table => new DASDatabricksTable(this, warehouse, table))
+    }
   }
+
+  def execute(stmt: ExecuteStatementRequest): DASDatabricksExecuteResult = {
+    val executeAPI = databricksClient.statementExecution()
+    val response1 = executeAPI.executeStatement(stmt)
+    val response = getResult(response1)
+    new DASDatabricksExecuteResult(executeAPI, response)
+  }
+
+  private val POLLING_TIME = 1000
+
+  @tailrec
+  private def getResult(response: StatementResponse): StatementResponse = {
+    val state = response.getStatus.getState
+    state match {
+      case StatementState.PENDING | StatementState.RUNNING =>
+        Thread.sleep(POLLING_TIME)
+        val response2 = databricksClient.statementExecution().getStatement(response.getStatementId)
+        getResult(response2)
+      case StatementState.SUCCEEDED => response
+      case StatementState.FAILED =>
+        throw new RuntimeException(s"Query failed: ${response.getStatus.getError.getMessage}")
+      case StatementState.CLOSED =>
+        throw new RuntimeException(s"Query closed: ${response.getStatus.getError.getMessage}")
+      case StatementState.CANCELED =>
+        throw new RuntimeException(s"Query canceled: ${response.getStatus.getError.getMessage}")
+    }
+  }
+
+  private def mapDatabricksException(t: Throwable): Option[RuntimeException] = {
+    t match {
+      case e: DatabricksError =>
+        e match {
+          case e: BadRequest =>
+            Some(new DASSdkInvalidArgumentException(e.getErrorInfo.asScala.map(_.getReason).mkString(", "), e))
+          case e: NotFound =>
+            Some(new DASSdkInvalidArgumentException(e.getErrorInfo.asScala.map(_.getReason).mkString(", "), e))
+          case e: PermissionDenied => Some(new DASSdkPermissionDeniedException(e.getMessage, e))
+          case e: Unauthenticated  => Some(new DASSdkUnauthenticatedException(e.getMessage, e))
+          case e =>
+            logger.warn("Unhandled Databricks error", e)
+            None
+        }
+      case e: DatabricksException =>
+        e.getCause match {
+          case _: UnknownHostException => Some(new DASSdkInvalidArgumentException("invalid host", e))
+          case _ =>
+            logger.warn("Unhandled Databricks exception", e)
+            None
+        }
+      case _ =>
+        logger.warn("Unhandled exception", t)
+        None
+    }
+  }
+
+  private def withDatabricksException[T](block: => T): T = {
+    try {
+      block
+    } catch {
+      case t: Throwable =>
+        throw mapDatabricksException(t).getOrElse(t)
+    }
+  }
+
 }
