+++ src/test/scala/com/rawlabs/das/databricks/DASDatabricksTest.scala
@@ -13,7 +13,9 @@
 package com.rawlabs.das.databricks
 
 import scala.jdk.CollectionConverters._
+
 import org.scalatest.funsuite.AnyFunSuite
+
 import com.rawlabs.das.sdk.{DASSdkInvalidArgumentException, DASSdkUnauthenticatedException}
 import com.rawlabs.protocol.das.v1.query.{Operator, Qual, SimpleQual}
 import com.rawlabs.protocol.das.v1.tables.{Row => ProtoRow}
@@ -70,8 +72,10 @@ class DASDatabricksTest extends AnyFunSuite with StrictLogging {
     }
   }
 
-  test("Should fail to register Databricks with an invalid warehouse option") {
+  ignore("Should fail to register Databricks with an invalid warehouse option") {
     val invalidOptions = options + ("warehouse" -> "invalid")
+    // This test is ignored because the Databricks SDK does not validate the warehouse ID
+    // at the time of registration.
     assertThrows[DASSdkInvalidArgumentException] {
       new DASDatabricks(invalidOptions)
     }
